feat: run 러닝 데이터 저장 기능 추가

+++ src/main/java/com/sixpack/dorundorun/feature/run/domain/RunSegmentData.java
@@ -12,7 +12,7 @@ public record RunSegmentData(
 	Double altitude,
 	Long distance,
 	Long pace,
-	Long speed,
+	Double speed,
 	Integer cadence
 ) {
 }
